Refactor UbersmithCore.xx to remote.ubersmith.xx
An instance make more sense and prevents mangling with metaclasses.
